@@ -1,6 +1,6 @@
 <?php
 /* ============================================================================
- * Copyright 2019 Zindex Software
+ * Copyright 2019-2020 Zindex Software
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,14 +21,9 @@ use ArrayObject;
 
 abstract class AbstractCacheHandler implements ICacheHandler
 {
-    /** @var null|ArrayObject */
-    protected $data = null;
-
-    /** @var bool */
-    protected $autoCommit = true;
-
-    /** @var bool */
-    protected $needsCommit = false;
+    protected ?ArrayObject $data = null;
+    protected bool $autoCommit = true;
+    protected bool $needsCommit = false;
 
     /**
      * @param bool $autoCommit
