@@ -1,6 +1,6 @@
 <?php
 /* ============================================================================
- * Copyright 2019 Zindex Software
+ * Copyright 2019-2020 Zindex Software
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,37 +19,21 @@ namespace Opis\FileSystem;
 
 class Context
 {
-    /** @var array */
-    protected $options;
-
-    /** @var bool */
-    protected $blocking = false;
-
-    /** @var int */
-    protected $readTimeout = 0;
-
-    /** @var int */
-    protected $readMode = STREAM_BUFFER_NONE;
-
-    /** @var int */
-    protected $readBuffer = 8192;
-
-    /** @var int */
-    protected $writeMode = STREAM_BUFFER_NONE;
-
-    /** @var int */
-    protected $writeBuffer = 8192;
-
-    /** @var string */
-    protected $protocol;
-
+    protected array $options;
+    protected bool $blocking = false;
+    protected int $readTimeout = 0;
+    protected int $readMode = STREAM_BUFFER_NONE;
+    protected int $readBuffer = 8192;
+    protected int $writeMode = STREAM_BUFFER_NONE;
+    protected int $writeBuffer = 8192;
+    protected string $protocol;
     /** @var resource|null */
     protected $context = null;
 
     /**
      * Context constructor.
      * @param string $protocol
-     * @param null $context
+     * @param null|resource $context
      * @param array $options
      */
     public function __construct(string $protocol, $context = null, array $options = [])
