@@ -1,6 +1,6 @@
 <?php
 /* ============================================================================
- * Copyright 2019 Zindex Software
+ * Copyright 2019-2020 Zindex Software
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,34 +17,31 @@
 
 namespace Opis\FileSystem\Directory;
 
-use Opis\FileSystem\File\IFileInfo;
+use Opis\FileSystem\File\FileInfo;
+use Opis\FileSystem\ProtocolInfo;
 use Opis\FileSystem\Handler\CachedHandler;
-use Opis\FileSystem\IProtocolInfo;
 use Opis\FileSystem\Traits\DirectoryFullPathTrait;
 
-final class CachedDirectory implements IDirectory
+final class CachedDirectory implements Directory
 {
     use DirectoryFullPathTrait;
 
-    /** @var \Opis\FileSystem\Directory\IDirectory */
-    private $directory;
-    /** @var CachedHandler */
-    private $handler;
-    /** @var string */
-    private $path;
+    private string $path;
+    private ?Directory $directory;
+    private ?CachedHandler $handler;
 
     /**
      * CachedDirectory constructor.
-     * @param IDirectory $directory
+     * @param Directory $directory
      * @param CachedHandler $handler
      */
-    public function __construct(IDirectory $directory, CachedHandler $handler)
+    public function __construct(Directory $directory, CachedHandler $handler)
     {
         $this->directory = $directory;
         $this->handler = $handler;
         $this->path = $directory->path();
 
-        if ($directory instanceof IProtocolInfo) {
+        if ($directory instanceof ProtocolInfo) {
             $this->protocol = $directory->protocol();
         }
     }
@@ -60,7 +57,7 @@ final class CachedDirectory implements IDirectory
     /**
      * @inheritDoc
      */
-    public function doNext(): ?IFileInfo
+    public function doNext(): ?FileInfo
     {
         if ($this->directory === null) {
             return null;
