@@ -1,6 +1,6 @@
 <?php
 /* ============================================================================
- * Copyright 2019 Zindex Software
+ * Copyright 2019-2020 Zindex Software
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,16 +17,14 @@
 
 namespace Opis\FileSystem;
 
-use Opis\Stream\Stream;
 use Opis\FileSystem\File\Stat;
+use Opis\Stream\ResourceStream;
 
-class FileStream extends Stream
+class FileStream extends ResourceStream
 {
     /** @var callable */
     protected $saveHandler = null;
-
-    /** @var null|array */
-    protected $stat = null;
+    protected ?array $stat = null;
 
     /**
      * FileStream constructor.
@@ -36,7 +34,8 @@ class FileStream extends Stream
      * @param callable|null $saveHandler
      * @param null|string $data
      */
-    public function __construct($stream, string $mode, ?Stat $stat = null, ?callable $saveHandler = null, ?string $data = null)
+    public function __construct($stream, string $mode, ?Stat $stat = null,
+        ?callable $saveHandler = null, ?string $data = null)
     {
         parent::__construct($stream, $mode);
 
