@@ -1,6 +1,6 @@
 <?php
 /* ============================================================================
- * Copyright 2019 Zindex Software
+ * Copyright 2019-2020 Zindex Software
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,23 +17,21 @@
 
 namespace Opis\FileSystem\Handler;
 
+use Throwable;
 use RecursiveDirectoryIterator, RecursiveIteratorIterator, FilesystemIterator;
-use Opis\Stream\{IStream, Printer\CopyPrinter};
+use Opis\Stream\{Stream, Printer\CopyPrinter};
 use Opis\FileSystem\FileStream;
 use Opis\FileSystem\Traits\SearchTrait;
-use Opis\FileSystem\File\{IFileInfo, FileInfo, Stat};
-use Opis\FileSystem\Directory\{IDirectory, LocalDirectory};
+use Opis\FileSystem\File\{FileInfo, Stat};
+use Opis\FileSystem\Directory\{Directory, LocalDirectory};
 
-class LocalFileHandler implements IFileSystemHandler, IAccessHandler, ISearchHandler
+class LocalFileHandler implements FileSystemHandler, AccessHandler, SearchHandler
 {
     use SearchTrait;
 
-    /** @var string */
-    protected $root;
-    /** @var string|null */
-    protected $baseUrl;
-    /** @var int */
-    protected $defaultMode = 0777;
+    protected string $root;
+    protected ?string $baseUrl = null;
+    protected int $defaultMode = 0777;
 
     /**
      * LocalFileHandler constructor.
@@ -70,7 +68,7 @@ class LocalFileHandler implements IFileSystemHandler, IAccessHandler, ISearchHan
     /**
      * @inheritDoc
      */
-    public function mkdir(string $path, int $mode = 0777, bool $recursive = true): ?IFileInfo
+    public function mkdir(string $path, int $mode = 0777, bool $recursive = true): ?FileInfo
     {
         $fullPath = $this->fullPath($path);
         if (is_dir($fullPath)) {
@@ -123,7 +121,7 @@ class LocalFileHandler implements IFileSystemHandler, IAccessHandler, ISearchHan
     /**
      * @inheritDoc
      */
-    public function touch(string $path, int $time, ?int $atime = null): ?IFileInfo
+    public function touch(string $path, int $time, ?int $atime = null): ?FileInfo
     {
         if (!$this->ensureDir($path, $this->defaultMode)) {
             return null;
@@ -139,7 +137,7 @@ class LocalFileHandler implements IFileSystemHandler, IAccessHandler, ISearchHan
     /**
      * @inheritDoc
      */
-    public function chmod(string $path, int $mode): ?IFileInfo
+    public function chmod(string $path, int $mode): ?FileInfo
     {
         if (!@chmod($this->fullPath($path), $mode)) {
             return null;
@@ -150,7 +148,7 @@ class LocalFileHandler implements IFileSystemHandler, IAccessHandler, ISearchHan
     /**
      * @inheritDoc
      */
-    public function chown(string $path, string $owner): ?IFileInfo
+    public function chown(string $path, string $owner): ?FileInfo
     {
         if (!@chown($this->fullPath($path), $owner)) {
             return null;
@@ -161,7 +159,7 @@ class LocalFileHandler implements IFileSystemHandler, IAccessHandler, ISearchHan
     /**
      * @inheritDoc
      */
-    public function chgrp(string $path, string $group): ?IFileInfo
+    public function chgrp(string $path, string $group): ?FileInfo
     {
         if (!@chgrp($this->fullPath($path), $group)) {
             return null;
@@ -172,7 +170,7 @@ class LocalFileHandler implements IFileSystemHandler, IAccessHandler, ISearchHan
     /**
      * @inheritDoc
      */
-    public function rename(string $from, string $to): ?IFileInfo
+    public function rename(string $from, string $to): ?FileInfo
     {
         if ($from === $to) {
             return null;
@@ -186,7 +184,7 @@ class LocalFileHandler implements IFileSystemHandler, IAccessHandler, ISearchHan
     /**
      * @inheritDoc
      */
-    public function copy(string $from, string $to, bool $overwrite = true): ?IFileInfo
+    public function copy(string $from, string $to, bool $overwrite = true): ?FileInfo
     {
         $from = trim($from, ' /');
         $to = trim($to, ' /');
@@ -261,7 +259,7 @@ class LocalFileHandler implements IFileSystemHandler, IAccessHandler, ISearchHan
     /**
      * @inheritDoc
      */
-    public function write(string $path, IStream $stream, int $mode = 0777): ?IFileInfo
+    public function write(string $path, Stream $stream, int $mode = 0777): ?FileInfo
     {
         if ($stream->size() === 0 && $stream->isEOF()) {
             $now = time();
@@ -287,7 +285,7 @@ class LocalFileHandler implements IFileSystemHandler, IAccessHandler, ISearchHan
     /**
      * @inheritDoc
      */
-    public function file(string $path, string $mode = 'rb'): ?IStream
+    public function file(string $path, string $mode = 'rb'): ?Stream
     {
         if ($stat = $this->stat($path)) {
             if ($stat->isDir()) {
@@ -301,7 +299,7 @@ class LocalFileHandler implements IFileSystemHandler, IAccessHandler, ISearchHan
 
         try {
             return new FileStream($this->fullPath($path), $mode, $stat);
-        } catch (\Throwable $e) {
+        } catch (Throwable $e) {
             return null;
         }
     }
@@ -309,7 +307,7 @@ class LocalFileHandler implements IFileSystemHandler, IAccessHandler, ISearchHan
     /**
      * @inheritDoc
      */
-    public function dir(string $path): ?IDirectory
+    public function dir(string $path): ?Directory
     {
         $stat = $this->stat($path);
         if (!$stat || !$stat->isDir()) {
@@ -321,7 +319,7 @@ class LocalFileHandler implements IFileSystemHandler, IAccessHandler, ISearchHan
     /**
      * @inheritDoc
      */
-    public function info(string $path): ?IFileInfo
+    public function info(string $path): ?FileInfo
     {
         $path = trim($path, ' /');
         if ($path === '') {
@@ -374,10 +372,10 @@ class LocalFileHandler implements IFileSystemHandler, IAccessHandler, ISearchHan
 
     /**
      * @param string $path
-     * @param IStream $stream
+     * @param Stream $stream
      * @return bool
      */
-    protected function writeFile(string $path, IStream $stream): bool
+    protected function writeFile(string $path, Stream $stream): bool
     {
         $path = $this->fullPath($path);
 
