@@ -1,6 +1,6 @@
 <?php
 /* ============================================================================
- * Copyright 2019 Zindex Software
+ * Copyright 2019-2020 Zindex Software
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,8 +19,7 @@ namespace Opis\FileSystem\Traits;
 
 trait ProtocolTrait
 {
-    /** @var string */
-    protected $protocol = '';
+    protected string $protocol = '';
 
     /**
      * @param string $path
