@@ -1,6 +1,6 @@
 <?php
 /* ============================================================================
- * Copyright 2019 Zindex Software
+ * Copyright 2019-2020 Zindex Software
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,12 +17,11 @@
 
 namespace Opis\FileSystem\Traits;
 
-use Opis\FileSystem\Directory\IDirectory;
+use Opis\FileSystem\Directory\Directory;
 
 trait StreamDirectoryTrait
 {
-    /** @var IDirectory|null */
-    protected $dir = null;
+    protected ?Directory $dir = null;
 
     /**
      * @inheritDoc
@@ -85,7 +84,7 @@ trait StreamDirectoryTrait
 
     /**
      * @param string $path
-     * @return IDirectory|null
+     * @return Directory|null
      */
-    abstract protected function dir(string $path): ?IDirectory;
+    abstract protected function dir(string $path): ?Directory;
 }
