@@ -1,6 +1,6 @@
 <?php
 /* ============================================================================
- * Copyright 2019 Zindex Software
+ * Copyright 2019-2020 Zindex Software
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,36 +15,14 @@
  * limitations under the License.
  * ============================================================================ */
 
-namespace Opis\FileSystem\Directory;
+namespace Opis\FileSystem;
 
-use Opis\FileSystem\File\IFileInfo;
-
-interface IDirectory
+interface FileSystemHandlerManager
 {
     /**
-     * @return string
+     * @param string $path
+     * @param string $protocol
+     * @return FileSystemStreamPathInfo|null
      */
-    public function path(): string;
-
-    /**
-     * @return string
-     */
-    public function fullPath(): string;
-
-    /**
-     * Next item in directory
-     * @return IFileInfo|null
-     */
-    public function next(): ?IFileInfo;
-
-    /**
-     * Goes back to first item
-     * @return bool
-     */
-    public function rewind(): bool;
-
-    /**
-     * Closes the handle
-     */
-    public function close(): void;
+    public function handle(string $path, string $protocol): ?FileSystemStreamPathInfo;
 }
