@@ -1,6 +1,6 @@
 <?php
 /* ============================================================================
- * Copyright 2019 Zindex Software
+ * Copyright 2019-2020 Zindex Software
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,30 +15,18 @@
  * limitations under the License.
  * ============================================================================ */
 
-namespace Opis\FileSystem\Traits;
+namespace Opis\FileSystem;
 
-use Opis\FileSystem\Context;
-
-trait ContextTrait
+interface ProtocolInfo
 {
-    /** @var null|Context */
-    protected $context = null;
+    /**
+     * @return string|null
+     */
+    public function protocol(): ?string;
 
     /**
-     * @inheritDoc
+     * @param string|null $protocol
+     * @return ProtocolInfo|static
      */
-    public function setContext(?Context $context): bool
-    {
-        $this->context = $context;
-
-        return true;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getContext(): ?Context
-    {
-        return $this->context;
-    }
+    public function setProtocol(?string $protocol): ProtocolInfo;
 }
